tests: isotp_fast: integrate conformance tests into single suite
