+++ tests/isotp_fast/conformance/prj.conf
@@ -1,10 +1,8 @@
 CONFIG_CAN=y
-CONFIG_CAN_FD_MODE=y
 CONFIG_ZTEST=y
 CONFIG_ZTEST_NEW_API=y
 CONFIG_ISOTP=y
 CONFIG_ISOTP_FAST=y
 CONFIG_ISOTP_USE_TX_BUF=y
 CONFIG_ISOTP_RX_BUF_COUNT=8
-CONFIG_ISOTP_FAST_RX_MAX_PACKET_COUNT=5
-CONFIG_ISOTP_FAST_PER_FRAME_DISPATCH=y
+CONFIG_ISOTP_FAST_RX_MAX_PACKET_COUNT=40
