+++ tests/isotp_fast/conformance/src/async_recv.h
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "../../conformance/src/main.h"
+#include "main.h"
 
 struct recv_msg
 {
