+++ tests/isotp_fast/conformance/testcase.yaml
@@ -1,0 +1,34 @@
+tests:
+  thingset_sdk.isotp_fast.conformance.sync.fd:
+    tags:
+      - can
+      - isotp
+    depends_on: can
+    integration_platforms:
+      - native_posix_64
+    filter: dt_chosen_enabled("zephyr,canbus") and not dt_compat_enabled("kvaser,pcican")
+    extra_args: EXTRA_CFLAGS=-Wno-error
+    extra_configs:
+      - CONFIG_ISOTP_FAST_BLOCKING_RECEIVE=y
+  thingset_sdk.isotp_fast.conformance.async:
+    tags:
+      - can
+      - isotp
+    depends_on: can
+    integration_platforms:
+      - native_posix_64
+    filter: dt_chosen_enabled("zephyr,canbus") and not dt_compat_enabled("kvaser,pcican")
+    extra_configs:
+      - CONFIG_ISOTP_FAST_PER_FRAME_DISPATCH=y
+  thingset_sdk.isotp_fast.conformance.async.fd:
+    tags:
+      - can
+      - isotp
+    depends_on: can
+    integration_platforms:
+      - native_posix_64
+    filter: dt_chosen_enabled("zephyr,canbus") and not dt_compat_enabled("kvaser,pcican")
+    extra_args: EXTRA_CFLAGS=-Wno-error
+    extra_configs:
+      - CONFIG_ISOTP_FAST_PER_FRAME_DISPATCH=y
+      - CONFIG_CAN_FD_MODE=y
