--- tests/isotp_fast/conformance_async/src/main.c
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2023 Brill Power
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#define ISOTP_FAST_CONFORMANCE_TEST_SUITE isotp_fast_conformance_async
-#include "async_recv.h"
-
-#include "../../conformance/src/main.c"
