--- tests/isotp_fast/conformance_async/testcase.yaml
@@ -1,9 +1,0 @@
-tests:
-  thingset_sdk.isotp_fast.conformance.async:
-    tags:
-      - can
-      - isotp
-    depends_on: can
-    integration_platforms:
-      - native_posix_64
-    filter: dt_chosen_enabled("zephyr,canbus") and not dt_compat_enabled("kvaser,pcican")
