--- tests/isotp_fast/conformance_sync/prj.conf
@@ -1,9 +1,0 @@
-CONFIG_CAN=y
-CONFIG_ZTEST=y
-CONFIG_ZTEST_NEW_API=y
-CONFIG_ISOTP=y
-CONFIG_ISOTP_FAST=y
-CONFIG_ISOTP_USE_TX_BUF=y
-CONFIG_ISOTP_RX_BUF_COUNT=8
-CONFIG_ISOTP_FAST_RX_MAX_PACKET_COUNT=40
-CONFIG_ISOTP_FAST_BLOCKING_RECEIVE=y
